+++ Packages/com.unity.inputsystem/Tests/InputSystem/CoreTests_Devices.cs
@@ -18,6 +18,9 @@ using Gyroscope = UnityEngine.Experimental.Input.Gyroscope;
 using UnityEngine.TestTools.Constraints;
 using Is = UnityEngine.TestTools.Constraints.Is;
 using Property = NUnit.Framework.PropertyAttribute;
+using Quaternion = UnityEngine.Quaternion;
+using Vector2 = UnityEngine.Vector2;
+using Vector3 = UnityEngine.Vector3;
 
 ////TODO: test that device re-creation doesn't lose flags and such
 
@@ -2193,10 +2196,55 @@ partial class CoreTests
 
         Assert.That(pen.inRange.ReadValue(), Is.EqualTo(0).Within(0.00001));
 
-        InputSystem.QueueStateEvent(pen, new PenState().WithButton(PenState.Button.InRange));
+        InputSystem.QueueStateEvent(pen, new PenState().WithButton(PenButton.InRange));
         InputSystem.Update();
 
         Assert.That(pen.inRange.ReadValue(), Is.EqualTo(1).Within(0.00001));
+    }
+
+    [Test]
+    [Category("Devices")]
+    public void Devices_CanUsePenAsPointer()
+    {
+        var pen = InputSystem.AddDevice<Pen>();
+
+        Assert.That(pen.phase.ReadValue(), Is.EqualTo(PointerPhase.None));
+        Assert.That(pen.position.ReadValue(), Is.EqualTo(Vector2.zero).Using(Vector2EqualityComparer.Instance));
+        Assert.That(pen.delta.ReadValue(), Is.EqualTo(Vector2.zero).Using(Vector2EqualityComparer.Instance));
+        Assert.That(pen.pressure.ReadValue(), Is.Zero);
+        Assert.That(pen.button.isPressed, Is.False);
+        Assert.That(pen.tip.isPressed, Is.False);
+        Assert.That(pen.eraser.isPressed, Is.False);
+        Assert.That(pen.firstBarrelButton.isPressed, Is.False);
+        Assert.That(pen.secondBarrelButton.isPressed, Is.False);
+        Assert.That(pen.thirdBarrelButton.isPressed, Is.False);
+        Assert.That(pen.fourthBarrelButton.isPressed, Is.False);
+        Assert.That(pen.inRange.isPressed, Is.False);
+        Assert.That(pen.twist.ReadValue(), Is.Zero);
+        Assert.That(pen.tilt.ReadValue(), Is.EqualTo(Vector2.zero).Using(Vector2EqualityComparer.Instance));
+
+        InputSystem.QueueStateEvent(pen, new PenState
+        {
+            position = new Vector2(0.123f, 0.234f),
+            delta = new Vector2(0.11f, 0.22f),
+            pressure = 0.345f,
+            twist = 0.456f,
+            tilt = new Vector2(0.567f, 0.678f),
+        });
+        InputSystem.Update();
+
+        Assert.That(pen.position.ReadValue(), Is.EqualTo(new Vector2(0.123f, 0.234f)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(pen.delta.ReadValue(), Is.EqualTo(new Vector2(0.11f, 0.22f)).Using(Vector2EqualityComparer.Instance));
+        Assert.That(pen.pressure.ReadValue(), Is.EqualTo(0.345).Within(0.00001));
+        Assert.That(pen.twist.ReadValue(), Is.EqualTo(0.456).Within(0.00001));
+        Assert.That(pen.tilt.ReadValue(), Is.EqualTo(new Vector2(0.567f, 0.678f)).Using(Vector2EqualityComparer.Instance));
+
+        AssertButtonPress(pen, new PenState().WithButton(PenButton.Tip), pen.tip, pen.button);
+        AssertButtonPress(pen, new PenState().WithButton(PenButton.Eraser), pen.eraser);
+        AssertButtonPress(pen, new PenState().WithButton(PenButton.BarrelFirst), pen.firstBarrelButton);
+        AssertButtonPress(pen, new PenState().WithButton(PenButton.BarrelSecond), pen.secondBarrelButton);
+        AssertButtonPress(pen, new PenState().WithButton(PenButton.BarrelThird), pen.thirdBarrelButton);
+        AssertButtonPress(pen, new PenState().WithButton(PenButton.BarrelFourth), pen.fourthBarrelButton);
     }
 
     ////FIXME: this needs to be overhauled; functioning of Touchscreen as Pointer is currently broken